Fix flow vlan parsing
The way the sed statement was setup, the first script was never passing
anything to the second so only matches from the first were being
recorded resulting in many vlans not being recorded as having flows.

Also records the actual flows associated with a vlan and stores them
in flowinfo/vlans/<vlan>/flows and creates a symlink
flowinfo/vlans/<vlan>/<vlan> back to the root vlan listing.