Prioritise search for mod_vlan_vid in flow vlan check
In other vlan ids that are part of mod_vlan_vid will be expected
to exist on a port attached to one of the bridges. This is of
course not always going to be true since this flow might be for
egress out of the switch to the external network and that is why
this check is only performed under --check-flow-vlans so that the
user can exclude vlans they know are external.