Fix mod_dl_src parsing
A typo in the code was stopping mac addresses
being paired with their local device. Results
are now presented slightly differently in that
for ingress substitutions, dst mac is shown
as a directory with contents for each
substituded src mac. For egress, the new mac
is presented as a dir with each substituted mac
as sub entries.